Starter-01/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
deepgram-sdk==2.3.0
python-dotenv==1.0.0